# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Video Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all generated videos
 */
export const ListVideosResponseItem = zod.object({
  id: zod.number(),
  topic: zod.string(),
  style: zod.string(),
  durationMinutes: zod.number(),
  voice: zod.string(),
  language: zod.string().optional(),
  status: zod
    .string()
    .describe(
      "pending | generating_script | generating_audio | generating_images | assembling_video | done | error",
    ),
  progress: zod.number().optional().describe("Progress percentage 0-100"),
  errorMessage: zod.string().optional(),
  outputPath: zod.string().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date().optional(),
});
export const ListVideosResponse = zod.array(ListVideosResponseItem);

/**
 * @summary Start a new video generation job
 */
export const createVideoBodyDurationMinutesMax = 30;

export const CreateVideoBody = zod.object({
  topic: zod.string().describe("The video topic"),
  style: zod
    .string()
    .describe("Video style (e.g. curious, mysterious, educational)"),
  durationMinutes: zod
    .number()
    .min(1)
    .max(createVideoBodyDurationMinutesMax)
    .describe("Target video duration in minutes"),
  voice: zod
    .string()
    .describe("Voice style (alloy, echo, fable, onyx, nova, shimmer)"),
  language: zod
    .string()
    .optional()
    .describe("Language for narration (default pt-BR)"),
});

/**
 * @summary Get a video by ID
 */
export const GetVideoParams = zod.object({
  id: zod.coerce.number(),
});

export const GetVideoResponse = zod.object({
  id: zod.number(),
  topic: zod.string(),
  style: zod.string(),
  durationMinutes: zod.number(),
  voice: zod.string(),
  language: zod.string().optional(),
  status: zod
    .string()
    .describe(
      "pending | generating_script | generating_audio | generating_images | assembling_video | done | error",
    ),
  progress: zod.number().optional().describe("Progress percentage 0-100"),
  errorMessage: zod.string().optional(),
  outputPath: zod.string().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Run the full video generation pipeline (SSE)
 */
export const GenerateVideoParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Download the final generated video
 */
export const DownloadVideoParams = zod.object({
  id: zod.coerce.number(),
});