# lib/api-zod/src/generated/types/createVideoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Video Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoBody {
  /** The video topic */
  topic: string;
  /** Video style (e.g. curious, mysterious, educational) */
  style: string;
  /**
   * Target video duration in minutes
   * @minimum 1
   * @maximum 30
   */
  durationMinutes: number;
  /** Voice style (alloy, echo, fable, onyx, nova, shimmer) */
  voice: string;
  /** Language for narration (default pt-BR) */
  language?: string;
}