# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Video Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  topic: string;
  style: string;
  durationMinutes: number;
  voice: string;
  language?: string;
  /** pending | generating_script | generating_audio | generating_images | assembling_video | done | error */
  status: string;
  /** Progress percentage 0-100 */
  progress?: number;
  errorMessage?: string;
  outputPath?: string;
  createdAt: Date;
  updatedAt?: Date;
}